backend/requirements.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2019.3
simplejson==3.17.0
sqlparse==0.3.0
